% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{ensembl_lookup}
\alias{ensembl_lookup}
\title{Look up a gene by Ensembl ID}
\usage{
ensembl_lookup(id)
}
\arguments{
\item{id}{Ensembl ID (e.g. "ENSG00000139618")}
}
\value{
tibble: same columns as ensembl_gene
}
\description{
Look up a gene by Ensembl ID
}
