% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{ensembl_sequence}
\alias{ensembl_sequence}
\title{Get gene sequence}
\usage{
ensembl_sequence(id, type = "genomic")
}
\arguments{
\item{id}{Ensembl ID}

\item{type}{"genomic" (default), "cdna", "cds", "protein"}
}
\value{
tibble: one row with id, seq (character), molecule_type
}
\description{
Get gene sequence
}
