% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.R
\name{ensembl_variant}
\alias{ensembl_variant}
\title{Look up variant by rsID}
\usage{
ensembl_variant(rsid, species = "homo_sapiens")
}
\arguments{
\item{rsid}{Variant ID (e.g. "rs699")}

\item{species}{Species (default "homo_sapiens")}
}
\value{
tibble: id, consequence_type, alleles, minor_allele, maf
}
\description{
Look up variant by rsID
}
